backend/requirements.txt
fastapi
uvicorn[standard]
spotipy
ytmusicapi
python-dotenv
httpx
fuzzywuzzy
python-Levenshtein
pydantic
requests
itsdangerous
